requirements.txt
django == 4.0.3

django-tinymce